arm64: dts: hi3798cv200: increase g-tx-fifo-size setting
The OTG device core has only two endpoints, which are working in fixed
direction, i.e. ep1 is IN (from device to host) and ep2 is OUT (from
host to device).  So there is no point to specify two numbers in
g-tx-fifo-size but just to waste fifo space.  Let's change it to 256
for full utilization of TX fifo.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>